backend/requirements.txt
aioredis==2.0.1
alembic==1.8.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
asyncpg==0.25.0
click==8.1.3
databases==0.6.0
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
fastapi-pagination==0.9.3
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pydantic==1.9.1
pyhumps==3.7.1
PyJWT==2.4.0
pyparsing==3.0.9
python-jose==3.3.0
redis==4.3.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
sse-starlette==0.10.3
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
wrapt==1.14.1
zipp==3.8.0
